adding position manager
